% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_code.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{Calculates BMI based on raw values}
\usage{
calculate_bmi(height, weight)
}
\arguments{
\item{height}{Height in meters}

\item{weight}{Weight in kilograms}
}
\value{
numeric value for the BMI
}
\description{
Calculates BMI based on raw values
}
\examples{
calculate_bmi(height = 1.70, weight = 65)
}
