% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_code.R
\name{calculate_bmi_df}
\alias{calculate_bmi_df}
\title{Calculates BMI in a data frame}
\usage{
calculate_bmi_df(df)
}
\arguments{
\item{df}{Data frame with at least two columns: height and weight}
}
\value{
Data frame (same as input) with an extra column for the 'bmi'
}
\description{
Calculates BMI in a data frame
}
\examples{
calculate_bmi_df(df = input_df)
}
